Test menos el de integracion

+++ backend/src/main/java/com/kiwi/features/quests/controllers/SubquestRepository.java
@@ -4,8 +4,8 @@ import com.kiwi.features.quests.data.SubquestPersistence;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface SubquestRepository extends JpaRepository<SubquestPersistence, Long> {
+public interface SubquestRepository extends JpaRepository<SubquestPersistence, Integer> {
 
-    List<SubquestPersistence> findAllByQuestIdOrderByOrderIndex(Long questId);
+    List<SubquestPersistence> findAllByQuestIdOrderByOrderIndex(int questId);
 
 }
