+++ backend/src/main/java/com/kiwi/features/quests/data/QuestDTO.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder @EqualsAndHashCode @ToString
 public class QuestDTO {
-    private Long questId;
+    private int questId;
     private String name;
     private String description;
     private int experience;
