+++ backend/src/main/java/com/kiwi/features/quests/data/QuestDomain.java
@@ -7,14 +7,14 @@ import java.util.List;
 @Setter
 public class QuestDomain {
 
-    private final Long questId;
+    private final int questId;
     private final String name;
     private final String description;
     private final int experience;
     private final QuestStatus status;
     private final List<SubquestDomain> subquests;
 
-    public QuestDomain(Long questId, String name, String description, int experience,
+    public QuestDomain(int questId, String name, String description, int experience,
                        QuestStatus status, List<SubquestDomain> subquests) {
         this.questId = questId;
         this.name = name;
