+++ backend/src/main/java/com/kiwi/features/quests/data/SubquestDomain.java
@@ -7,13 +7,13 @@ import lombok.Setter;
 @Setter
 public class SubquestDomain {
 
-    private final Long subquestId;
+    private final int subquestId;
     private final String name;
     private final int experience;
     private final int order;
     private final SubquestStatus status;
 
-    public SubquestDomain(Long subquestId, String name, int experience, int order, SubquestStatus status) {
+    public SubquestDomain(int subquestId, String name, int experience, int order, SubquestStatus status) {
         this.subquestId = subquestId;
         this.name = name;
         this.experience = experience;
