+++ backend/src/main/java/com/kiwi/features/quests/exceptions/QuestNotFoundException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.quests.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class QuestNotFoundException extends RuntimeException {
-    public QuestNotFoundException(@NotNull long id) {
+    public QuestNotFoundException(@NotNull int id) {
         super(String.format("Cannot found quest with id %s", id));
     }
 }
