+++ backend/src/main/java/com/kiwi/features/quests/exceptions/SubquestNotFoundException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.quests.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class SubquestNotFoundException extends RuntimeException {
-    public SubquestNotFoundException(@NotNull long id) {
+    public SubquestNotFoundException(@NotNull int id) {
         super(String.format("Cannot found subquest with id %s", id));
     }
 }
