+++ backend/src/main/java/com/kiwi/features/quests/exceptions/SubquestStatusNotFoundException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.quests.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class SubquestStatusNotFoundException extends RuntimeException {
-    public SubquestStatusNotFoundException(@NotNull long id) {
+    public SubquestStatusNotFoundException(@NotNull int id) {
         super(String.format("Cannot found status of subquest with id %s", id));
     }
 }
